travail sur view/panier

+++ laravel/app/Http/Controllers/ForfaitController.php
@@ -106,17 +106,46 @@ class ForfaitController extends Controller
     // }
 
 
+    // public function destroy($id)
+    // {
+    //     $userId = auth()->id();
+
+    //     if ($userId) {
+    //         DB::table('user_forfait')->where('id', $id)->delete();
+    //         return redirect()->route('user.index')->with('success', "L'entrée de la table user_forfait a été supprimée!");
+    //     }
+
+    //     return redirect()->route('user.index')->with('error', "La suppression de l'entrée de la table user_forfait a échoué!");
+    // }
+
+
+
+    // a verifier
     public function destroy($id)
     {
         $userId = auth()->id();
 
         if ($userId) {
-            DB::table('user_forfait')->where('id', $id)->delete();
-            return redirect()->route('user.index')->with('success', "L'entrée de la table user_forfait a été supprimée!");
+            $userForfait = DB::table('user_forfait')->find($id);
+
+            if ($userForfait) {
+                $dateArrivee = $userForfait->date_arrivee;
+                $dateMaxEvenement = '2024-08-11';
+
+                if (strtotime($dateArrivee) > strtotime($dateMaxEvenement)) {
+                    return redirect()->route('user.index')->with('error', "La date est dépassée. La suppression n'est pas autorisée.");
+                }
+
+                DB::table('user_forfait')->where('id', $id)->delete();
+                return redirect()->route('user.index')->with('success', "L'entrée de la table user_forfait a été supprimée!");
+            }
         }
 
         return redirect()->route('user.index')->with('error', "La suppression de l'entrée de la table user_forfait a échoué!");
     }
+
+
+
 
     public function destroyForfaitAdmin($id)
     {
